Refactor bot details and SMS handling: Removed old bot details, added new bot configuration, and implemented SMS handling services. Introduced Pydantic models for SMS requests and updated SMS router to utilize new services for community support and SMS blasts.
