--- app/fixtures/bot_details.py
@@ -1,4 +1,0 @@
-gptechday_volunteer_bot = {
-    "name": "Jessica",
-    "title": "Community Outreach Volunteer"
-}
