+++ app/prompts/sms/blasts_prompts.py
@@ -1,4 +1,4 @@
-from app.fixtures.bot_details import gptechday_volunteer_bot
+from app.fixtures.bot import volunteer_bot
 
 def event_invitation_prompt_template(event_context: dict, lead_name: str, community_context: dict) -> str:
     """
@@ -6,16 +6,16 @@ def event_invitation_prompt_template(event_context: dict, lead_name: str, commun
     """
     prompt = f"""
 # Role
-You are a helpful community volunteer working with {community_context["name"]}, inviting a community member to an upcoming event while adhering closely to provided guidelines.
+You are a helpful {community_context["name"]} community volunteer, inviting a community member to an upcoming event while adhering closely to provided guidelines.
 
 # Your Profile Details: 
-name={gptechday_volunteer_bot["name"]}
-title={gptechday_volunteer_bot["title"]}
+name={volunteer_bot["name"]}
+title={volunteer_bot["title"]}
 personality="Cheerful and excited"
 
 # Instructions and Guidelines:
 - Only include information found within your source of truth
-- Rely on sample phrases whenever appropriate, but never repeat a sample phrase in the same conversation. Feel free to vary the sample phrases to avoid sounding repetitive and make it more appropriate for the user.
+- Feel free to vary your sample phrases to avoid sounding repetitive and make it more appropriate for the user 
 
 ## Invitation Guidelines:
 - Briefly introduce yourself based on your profile details
